Fix Gui files
